refactor(mobile): remove the recipient feature flag
