+++ apps/mobile/src/features/send/components/recipient/recipient-selector/recipient-selector.tsx
@@ -1,18 +1,18 @@
 import { useState } from 'react';
 
-import { RecipientInput } from '@/features/send/components/recipient/v2/recipient-input';
-import { RecipientSelectorHeader } from '@/features/send/components/recipient/v2/recipient-selector/recipient-selector-header';
-import { RecipientSelectorItem } from '@/features/send/components/recipient/v2/recipient-selector/recipient-selector-item';
-import { RecipientSelectorSearchEmptyState } from '@/features/send/components/recipient/v2/recipient-selector/recipient-selector-search-empty-state';
-import { RecipientSelectorSectionHeader } from '@/features/send/components/recipient/v2/recipient-selector/recipient-selector-section-header';
+import { RecipientInput } from '@/features/send/components/recipient/recipient-input';
+import { RecipientSelectorHeader } from '@/features/send/components/recipient/recipient-selector/recipient-selector-header';
+import { RecipientSelectorItem } from '@/features/send/components/recipient/recipient-selector/recipient-selector-item';
+import { RecipientSelectorSearchEmptyState } from '@/features/send/components/recipient/recipient-selector/recipient-selector-search-empty-state';
+import { RecipientSelectorSectionHeader } from '@/features/send/components/recipient/recipient-selector/recipient-selector-section-header';
 import {
   isBnsLookupCandidate,
   normalizeSearchTerm,
-} from '@/features/send/components/recipient/v2/recipient.utils';
+} from '@/features/send/components/recipient/recipient.utils';
 import {
   matchSuggestionsResult,
   useRecipientSuggestions,
-} from '@/features/send/components/recipient/v2/use-recipient-suggestions';
+} from '@/features/send/components/recipient/use-recipient-suggestions';
 import { SendFormLoadingSpinner } from '@/features/send/components/send-form-layout';
 import { type Account } from '@/store/accounts/accounts';
 import { BottomSheetSectionList } from '@gorhom/bottom-sheet';
