+++ apps/mobile/src/features/send/forms/btc/btc-form.tsx
@@ -94,19 +94,11 @@ export function BtcForm({
       <Controller
         control={form.control}
         name="recipient"
-        render={({
-          field: { value, onChange, onBlur },
-          fieldState: { invalid, isDirty, isTouched, error },
-        }) => (
+        render={({ field: { value, onChange } }) => (
           <Recipient
             assetInfo={assetInfo}
             value={value}
             onChange={onChange}
-            onBlur={onBlur}
-            invalid={invalid}
-            isDirty={isDirty}
-            isTouched={isTouched}
-            error={error}
             recipientSchema={schema.sourceType().shape.recipient}
           />
         )}
